% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_palette.R
\name{extract_palette}
\alias{extract_palette}
\title{Extract palette from a ggnested object}
\usage{
extract_palette(ggnested_obj)
}
\arguments{
\item{ggnested_obj}{A ggplot2 object generated by ggnested}
}
\value{
A dataframe with the group_subgroup label, the main group colours
 and the subgroup colours.
}
\description{
This function extracts the colour palette used in a ggnested object.
}
