% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_palette.R
\name{nested_palette}
\alias{nested_palette}
\title{Generate a nested palette}
\usage{
nested_palette(
  data,
  group,
  subgroup,
  gradient_type = c("both", "shades", "tints"),
  min_l = 0.05,
  max_l = 0.95,
  palette = NULL,
  base_clr = "#008CF0",
  join_str = "_",
  min_c = 50,
  max_c = 90,
  power = 1
)
}
\arguments{
\item{data}{A dataframe with at least two columns}

\item{group}{Name of the main grouping variable in the dataframe}

\item{subgroup}{Name of the subgrouping variable in the dataframe}

\item{gradient_type}{whether to generate shades, tints, or both for each main
colour}

\item{min_l}{the lowest lightness value between 0 and 1 of each gradient if 
\code{type \%in\% c('both', 'shades')}, where 0 equals black and 1 equals white.}

\item{max_l}{the highest lightness value between 0 and 1 of each gradient if 
\code{type \%in\% c('both', 'tints')}, where 0 equals black and 1 equals white.}

\item{palette}{An optional palette of at least as many colours as there are 
unique values in the grouping variable. If named, names are assumed to corresponds
to \code{main_group} levels.}

\item{base_clr}{A colour from which to generate other colours if no palette 
is provided}

\item{join_str}{The string used to join the group and subgroup variables}

\item{min_c}{the lowest chroma value between 0 and 360 of each gradient if 
\code{type == 'fancy'}.}

\item{max_c}{the highest chroma value between 0 and 360 of each gradient if 
\code{type == 'fancy'}.}

\item{power}{control parameter determining how chroma and luminance need to be 
increased (1=linear, 2= quadratic, etc.) if \code{type == 'fancy'}.}
}
\value{
A dataframe with each group and its colour, each subgroup with its 
colour and the joined group_subgroup.
}
\description{
This function generates colours based on a grouping, and shades and tints 
of each colour based on a subgrouping.

 @details
 The 'fancy' gradient is generated using the \code{colorspace} package and needs
 fine tuningas, some parameter combinations do not yield a monochromal gradient.
}
