% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnested.R
\name{theme_nested}
\alias{theme_nested}
\title{Add a ggplot2 theme to a ggnested plot}
\usage{
theme_nested(theme_fun = theme, ...)
}
\arguments{
\item{theme_fun}{The theme* function to apply}

\item{...}{Arguments to be passed to the theme* function}
}
\value{
A ggplot2 theme
}
\description{
This is a wrapper function around any  ggplot2 theme* function to modify the 
theme of a ggnested plot.
}
