% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnested.R
\name{ggnested}
\alias{ggnested}
\title{Generate a ggplot2 object with a nested colour/fill structure}
\usage{
ggnested(
  data,
  mapping = aes(),
  ...,
  legend_labeling = c("sub", "join", "main"),
  join_str = " - ",
  legend_title = NULL,
  main_keys = TRUE,
  nested_aes = c("fill", "color"),
  gradient_type = c("both", "shades", "tints"),
  min_l = 0.05,
  max_l = 0.95,
  main_palette = NULL,
  base_clr = "#008CF0"
)
}
\arguments{
\item{data}{A dataframe}

\item{mapping}{Aesthetic mapping for the ggplot2 object. Specify both group 
and subgroup within the aesthetic mapping.}

\item{...}{Arguments to be passed to the ggplot function}

\item{legend_labeling}{Labeling for legend keys. \code{main} displays the 
label of the main group; \code{sub} displays the labels of the subgroups; 
\code{join} displays the joined labels of the main and subgroups, separated 
by \code{join_str}.}

\item{join_str}{String used to join main and subgroup labels}

\item{legend_title}{Legend title}

\item{main_keys}{Logical. Display main groups as titles within the legend?}

\item{nested_aes}{The aethetics in c("fill", "colour") to which to apply the nested
palette.}

\item{gradient_type}{whether to generate shades, tints, or both for each main
colour}

\item{min_l}{the lowest lightness value between 0 and 1 of each gradient if 
\code{type \%in\% c('both', 'shades')}, where 0 equals black and 1 equals white.}

\item{max_l}{the highest lightness value between 0 and 1 of each gradient if 
\code{type \%in\% c('both', 'tints')}, where 0 equals black and 1 equals white.}

\item{main_palette}{An optional palette of at least as many colours as there
are unique values in the main grouping variable. Names will be used to assign
colours if provided.}

\item{base_clr}{A colour from which to generate other colours if no palette 
is provided}

\item{palette}{An optional palette of at least as many colours as there are}
}
\value{
A ggplot2 object
}
\description{
This function generates a ggplot2 object for a dataframe with a nested or 
clustered structure. The main group or cluster is assigned a colour,
 and each subgroup within each main group is assigned a shade or tint of that 
 colour.
}
