% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_palette.R
\name{nested_palette}
\alias{nested_palette}
\title{Generate a nested palette}
\usage{
nested_palette(
  data,
  group,
  subgroup,
  palette = NULL,
  base_clr = "#008CF0",
  join_str = "_"
)
}
\arguments{
\item{data}{A dataframe with at least two columns}

\item{group}{Name of the main grouping variable in the dataframe}

\item{subgroup}{Name of the subgrouping variable in the dataframe}

\item{palette}{An optional palette of at least as many colours as there are 
unique values in the grouping variable}

\item{base_clr}{A colour from which to generate other colours if no palette 
is provided}

\item{join_str}{The string used to join the group and subgroup variables}
}
\value{
A dataframe with each group and its colour, each subgroup with its 
colour and the joined group_subgroup.
}
\description{
This function generates colours based on a grouping, and shades and tints 
of each colour based on a subgrouping.
}
